+++ src/Symfony/Component/Messenger/CHANGELOG.md
@@ -32,6 +32,7 @@ CHANGELOG
  * Add `WrappedExceptionsInterface` interface for exceptions that hold multiple individual exceptions
  * Deprecate `HandlerFailedException::getNestedExceptions()`, `HandlerFailedException::getNestedExceptionsOfClass()`
    and `DelayedMessageHandlingException::getExceptions()` which are replaced by a new `getWrappedExceptions()` method
+ * Separate `WorkerRunningEvent` into `WorkerBusyEvent` and `WorkerIdleEvent`
 
 6.3
 ---